chore(ops): clean up duplicates after ops refactor (#97)
